refactor: 메인화면 로직 컴포넌트로 분리
