+++ client/src/pages/home/GuideModeButton.tsx
@@ -1,0 +1,16 @@
+function GuideModeButton() {
+  return (
+    <div className="flex justify-center mb-[204px]">
+      <div className="flex justify-center gap-[80px] bg-grey-001 w-[533px] h-[90px] items-center rounded-[6px]">
+        <p className="title5 text-grey-black">
+          무엇을 골라야 할 지 모르겠다면?
+        </p>
+        <p className="text-[32px] font-hsans-head text-grey-black font-medium tracking-[-0.96px]">
+          Guide Mode
+        </p>
+      </div>
+    </div>
+  );
+}
+
+export default GuideModeButton;
