+++ client/src/pages/home/MakingCarButtonsBox.tsx
@@ -1,0 +1,20 @@
+function MakingCarButtonsBox() {
+  return (
+    <div className="flex justify-between px-[48px] mb-[69px]">
+      <div className="w-[140px] h-[50px] bg-main-blue text-white rounded-[6px] flex justify-center items-center">
+        내 차 만들기
+      </div>
+      <div className="w-[140px] h-[50px] bg-main-blue text-white rounded-[6px] flex justify-center items-center">
+        내 차 만들기
+      </div>
+      <div className="w-[140px] h-[50px] bg-main-blue text-white rounded-[6px] flex justify-center items-center">
+        내 차 만들기
+      </div>
+      <div className="w-[140px] h-[50px] bg-main-blue text-white rounded-[6px] flex justify-center items-center">
+        내 차 만들기
+      </div>
+    </div>
+  );
+}
+
+export default MakingCarButtonsBox;
