+++ client/src/pages/home/TrimCardsBox.tsx
@@ -1,0 +1,23 @@
+function TrimCarsBox() {
+  return (
+    <div className="flex justify-between gap-[16px]">
+      <div className="bg-white w-[192px] h-[123px] shrink-0 flex justify-center items-center">
+        trim card
+      </div>
+      <div className="bg-white w-[192px] h-[123px] shrink-0 flex justify-center items-center">
+        trim card
+      </div>
+      <div className="bg-white w-[192px] h-[123px] shrink-0 flex justify-center items-center">
+        trim card
+      </div>
+      <div className="bg-white w-[192px] h-[123px] shrink-0 flex justify-center items-center">
+        trim card
+      </div>
+      <div className="bg-white w-[192px] h-[123px] shrink-0 flex justify-center items-center">
+        trim card
+      </div>
+    </div>
+  );
+}
+
+export default TrimCarsBox;
